+++ src/renderer/model-apps/obj-detection-model/ObjDetectionModelInfo.tsx
@@ -1,0 +1,10 @@
+import { ModelAppConfig } from 'src/shared/models';
+import TemplateInfo from '../template/TemplateInfo';
+
+export default function ObjDetectionModelInfo({
+  modelAppConfig,
+}: {
+  modelAppConfig: ModelAppConfig;
+}) {
+  return <TemplateInfo modelAppConfig={modelAppConfig} />;
+}
